requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
sqlalchemy
requests
python-multipart
Pillow